refactor: StartPage 분리
